+++ How to/Customization  Print Quality/angular.json
@@ -96,6 +96,6 @@
     }
   },
   "cli": {
-    "analytics": "1ae3d0d1-c71a-4661-b93c-819a70da741e"
+    "analytics": "8d9c65e2-01c9-4a66-bc68-685c535cc647"
   }
 }
