+++ How to/Customization  Print Quality/src/index.html
@@ -8,7 +8,6 @@
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
 <body>
-  <br><br><br>
   <app-root></app-root>
 </body>
 </html>
