+++ How to/Customization  Print Quality/tsconfig.app.json
