+++ How to/Customization  Print Quality/tsconfig.json
